+++ click_extended/decorators/check/__init__.py
@@ -1,7 +1,9 @@
 """Initialization file for the `click_extended.decorators.check` module."""
 
 from click_extended.decorators.check.exclusive import exclusive
+from click_extended.decorators.check.min_length import min_length
 
 __all__ = [
     "exclusive",
+    "min_length",
 ]