+++ CHANGELOG.md
@@ -10,6 +10,8 @@
 - **Code Snippets**: Python code snippets.
 - **Dependencies**: Added `PyYAML` and `tomli` (only before 3.11) as a project dependency.
 - **Dev Dependencies**: Added `types-PyYAML` as a project dev dependency.
+- **`@min_length`**: Child decorator to enforce a minimum length of a string.
+- **`@max_length`**: Child decorator to enforce a maximum length of a string.
 - **`@less_than`**: Child decorator to check is a value is less or equal (inclusive=True) to a threshold.
 - **`@greater_than`**: Child decorator to check is a value is greater or equal (inclusive=True) to a threshold.
 - **`@to_timestamp`**: Child decorator to convert a `date`, `time`, or `datetime` object to a unix timestamp.